register broken but login and sstorage work
