+++ api/database.json
@@ -4,6 +4,11 @@
       "id": 1,
       "username": "mjb",
       "password": "mjb"
+    },
+    {
+      "username": "Jordan",
+      "password": "Jordan",
+      "id": 2
     }
   ],
   "chores": [
@@ -20,20 +25,6 @@
       "materials": "paper towels, windex",
       "date": "2020-12-15",
       "userId": 1
-    },
-    {
-      "activity": "Let the Dog Out",
-      "materials": "Stay in view of dog while she is outside",
-      "date": "2019-08-03",
-      "userId": "",
-      "id": 3
-    },
-    {
-      "activity": "Is this the edit form?",
-      "materials": "hands and brains",
-      "date": "1221-12-21",
-      "userId": 0,
-      "id": 4
     }
   ]
 }
